refactor(js): combine set & effect pattern files
