--- js/_hero-set.js
@@ -1,10 +1,0 @@
-export function set (el, pattern) {
-  el.style.backgroundImage = pattern
-  el.style.backgroundPosition = 'center'
-}
-
-export function setRandom (el, arr) {
-  let rand = Math.floor(Math.random() * arr.length)
-  let background = arr[rand]
-  set(el, background)
-}
