--- js/_images.js
@@ -1,7 +1,0 @@
-import Layzr from 'layzr.js/src/layzr.js'
-
-const lazy = Layzr({
-  normal: 'data-src'
-})
-
-export default lazy
